@@ -18,19 +18,20 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0-8.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.17.0"
 
 dependencies:
-  flutter_hooks: ^0.15.0
+  flutter_hooks: ^0.16.0-nullsafety.0
   flutter_layout_grid: ^0.11.0-nullsafety.1
-  flutter_svg: ^0.20.0-nullsafety
+  flutter_svg: ^0.22.0
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  mobx: ^1.2.1
-  flutter_mobx: ^1.1.0
-  provider: ^4.3.3
-  shared_preferences: ^0.5.12+4
+  http: ^0.13.1
+  mobx: ^2.0.1
+  flutter_mobx: ^2.0.0
+  provider: ^5.0.0
+  shared_preferences: ^2.0.5
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.[1,2].forEach((i) => i);
@@ -40,8 +41,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   pedantic: ^1.10.0-nullsafety
-  build_runner: ^1.10.9
-  mobx_codegen: ^1.1.2
+  build_runner: ^2.0.1
+  mobx_codegen: ^2.0.1+3
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
