@@ -11,7 +11,7 @@ import 'package:rulegamemobile/widgets/shape.dart';
 class Bucket extends HookWidget {
   final int pos;
 
-  Bucket({this.pos});
+  Bucket({required this.pos});
 
   @override
   Widget build(BuildContext context) {
@@ -19,10 +19,11 @@ class Bucket extends HookWidget {
     final hover = useState(false);
 
     return Container(
+      constraints: BoxConstraints.expand(),
       child: DragTarget<BoardObject>(
         onAccept: (boardObject) {
           hover.value = false;
-          return boardStore.move(boardObject.id, pos);
+          boardStore.move(boardObject.id, pos);
         },
         onWillAccept: (_) => true,
         onMove: (_) {
@@ -31,15 +32,15 @@ class Bucket extends HookWidget {
         onLeave: (_) {
           hover.value = false;
         },
-        builder: (context, List<Object> candidateData, rejectedData) =>
+        builder: (context, List<BoardObject?> candidateData, rejectedData) =>
             Transform.scale(
           scale: hover.value ? 4.0 : 1.0,
           child: Observer(
             builder: (_) => Shape(
-              shape: boardStore.bucketShapes[pos],
-              color: boardStore.bucketShapes[pos] == SpecialShape.BUCKET
-                  ? BUCKET_COLOR
-                  : null,
+              shape: boardStore.bucketShapes[pos]!,
+              // color: boardStore.bucketShapes[pos] == SpecialShape.BUCKET
+              //     ? BUCKET_COLOR
+              //     : null,
               opacity: (boardStore.isPaused &&
                           boardStore.bucketShapes[pos] ==
                               SpecialShape.BUCKET) ||
@@ -50,7 +51,6 @@ class Bucket extends HookWidget {
           ),
         ),
       ),
-      constraints: BoxConstraints.expand(),
     );
   }
 }
